refactor disease.js
