+++ static/js/disease/overview/checkSummaryData.js
@@ -1,0 +1,28 @@
+export function checkSummaryData(entryData) {
+  const summaryWrapper = document.querySelector('.summary-wrapper');
+  const summaryNav = document.querySelector('.nav-link.overview');
+  const diseaseDefinition = document.getElementById('temp-disease-definition');
+  const navBorderTop = document.querySelector(
+    '#temp-side-navigation > ul > li:first-child'
+  );
+  if (
+    !entryData.alt_label_ja &&
+    !entryData.alt_label_en &&
+    !entryData.db_xrefs?.omim &&
+    !entryData.db_xrefs?.orphanet &&
+    !entryData.medgen_id &&
+    !entryData.mondos &&
+    !entryData.kegg &&
+    !entryData.urdbms
+  ) {
+    if (summaryWrapper) {
+      summaryWrapper.style = 'display: none;';
+      summaryNav.style = 'display: none;';
+      navBorderTop.style = 'border-top: none;';
+      if (diseaseDefinition) {
+        summaryNav.style = 'display: block;';
+        navBorderTop.style = 'border-top: block;';
+      }
+    }
+  }
+}
