+++ static/js/disease/overview/makeAlternativeName.js
@@ -1,0 +1,34 @@
+export function makeAlternativeName(entryData) {
+  const altLabelJa = document.querySelector('.alt-label-ja');
+  const altLabelEn = document.querySelector('.alt-label-en');
+  const currentLang = document.querySelector('.language-select').value;
+
+  if (entryData.alt_label_ja && currentLang === 'ja') {
+    const divElement = document.createElement('div');
+    altLabelJa.append(divElement);
+    entryData.alt_label_ja.forEach((item) => {
+      const ddElement = document.createElement('dd');
+      ddElement.textContent = item;
+      ddElement.classList.add('linked-item', '-unlinked');
+      divElement.append(ddElement);
+    });
+  } else {
+    altLabelJa.remove();
+  }
+  if (entryData.alt_label_en) {
+    const divElement = document.createElement('div');
+    altLabelEn.append(divElement);
+    entryData.alt_label_en.forEach((item) => {
+      const ddElement = document.createElement('dd');
+      ddElement.textContent = item;
+      ddElement.classList.add('linked-item', '-unlinked');
+      divElement.append(ddElement);
+    });
+  } else {
+    altLabelEn.remove();
+  }
+
+  if (!(entryData.alt_label_ja || entryData.alt_label_en)) {
+    document.querySelector('.sub-title.-alt-label').remove();
+  }
+}
