+++ static/js/disease/overview/makeDiseaseDefinition.js
@@ -1,0 +1,65 @@
+export function makeDiseaseDefinition(entryData) {
+  const diseaseDefinition = document.getElementById('temp-disease-definition');
+  const tabWrap = diseaseDefinition.querySelector(
+    '#temp-disease-definition .tab-wrap'
+  );
+  const currentLang = document.querySelector('.language-select').value;
+
+  const items = [
+    {
+      class: 'mhlw',
+      existing: currentLang === 'en' ? false : !!entryData.description,
+      desc: entryData.description,
+      translate: false,
+    },
+    {
+      class: 'monarch-initiative',
+      existing: !!entryData.mondo_decs,
+      desc: entryData.mondo_decs?.map((dec) => dec.id).join(' '),
+      translate: true,
+    },
+    {
+      class: 'medgen',
+      existing: !!entryData.medgen_definition,
+      desc: entryData.medgen_definition,
+      translate: true,
+    },
+  ];
+
+  if (items.every((item) => !item.existing)) {
+    diseaseDefinition.remove();
+    document.querySelector('.sub-title.-disease-definition').remove();
+  } else {
+    let isFirstTab = true;
+
+    items.forEach((item) => {
+      if (!item.existing) {
+        const input = document.getElementById(`disease-${item.class}`);
+        const label = input.nextElementSibling;
+        label.remove();
+        input.remove();
+      }
+
+      const content = tabWrap.querySelector(`.${item.class}`);
+
+      content.textContent = item.desc;
+      const currentTab = tabWrap.querySelector(`#disease-${item.class}`);
+
+      if (currentTab && isFirstTab) {
+        currentTab.checked = true;
+        isFirstTab = false;
+      }
+
+      if (item.translate) {
+        const translationLink = document.createElement('a');
+        const translationUrl = `https://translate.google.co.jp/?hl=ja#en/ja/${item.desc}`;
+        translationLink.setAttribute('href', translationUrl);
+        translationLink.setAttribute('target', '_blank');
+        translationLink.setAttribute('rel', 'noopener noreferrer');
+        translationLink.innerHTML =
+          '<span class="google-translate">&nbsp;&gt;&gt;&nbsp;翻訳 (Google)</span>';
+        content.append(translationLink);
+      }
+    });
+  }
+}
