+++ static/js/disease/overview/makeExternalLinks.js
@@ -1,0 +1,35 @@
+export function makeExternalLinks(entryData) {
+  const externalLinks = document.getElementById('temp-external-links');
+
+  const items = [
+    {
+      url: entryData.mhlw?.url,
+      element: externalLinks.querySelector('.linked-item.mhlw'),
+      existing: !!entryData.mhlw,
+    },
+    {
+      url: entryData.source,
+      element: externalLinks.querySelector('.linked-item.source'),
+      existing: !!entryData.source,
+    },
+    {
+      url: entryData.nanbyou?.url,
+      element: externalLinks.querySelector('.linked-item.nanbyou'),
+      existing: !!entryData.nanbyou,
+    },
+    {
+      url: entryData.shouman?.url,
+      element: externalLinks.querySelector('.linked-item.shouman'),
+      existing: !!entryData.shouman,
+    },
+  ];
+
+  items.forEach((item) => {
+    const { url, element } = item;
+    if (url) {
+      element.querySelector('a').setAttribute('href', url);
+    } else {
+      element.remove();
+    }
+  });
+}
