+++ static/js/disease/overview/makeHeader.js
@@ -1,0 +1,46 @@
+export function makeHeader(entryData) {
+  const nandoId = entryData.nandoId;
+  const refNandoId = document.getElementById('temp-nando-id');
+  refNandoId.textContent = nandoId;
+
+  const refNandoLink = document.getElementById('temp-nando-link');
+  refNandoLink.setAttribute(
+    'href',
+    refNandoLink.getAttribute('href') + nandoId
+  );
+  refNandoLink.textContent += nandoId;
+
+  document
+    .getElementById('temp-nando-copy')
+    .addEventListener('click', async () => {
+      const clipboardText = 'https://nanbyodata.jp/ontology/NANDO_' + nandoId;
+      document.getElementById('temp-nando-copy').textContent = 'Copied!';
+      await navigator.clipboard.writeText(clipboardText);
+      await new Promise((resolve) => setTimeout(resolve, 1000));
+      document.getElementById('temp-nando-copy').textContent = 'Copy';
+    });
+
+  const labelJa = document.getElementById('temp-label-ja');
+  if (labelJa) {
+    labelJa.innerHTML =
+      '<ruby>' +
+      entryData.label_ja +
+      '<rt>' +
+      entryData.ruby +
+      '</rt>' +
+      '</ruby>';
+  }
+
+  const labelEn = document.getElementById('temp-label-en');
+  labelEn.textContent = entryData.label_en;
+
+  const notificationNumber = document.getElementById(
+    'temp-notification-number'
+  );
+  notificationNumber.textContent = entryData.notification_number;
+  if (!entryData.notification_number) {
+    notificationNumber.parentNode.remove();
+    const tempDataSummary = document.getElementById('overview');
+    tempDataSummary.style.borderBottom = 'none';
+  }
+}
