+++ static/js/utils/smart_box.js
